polis-optimization/src/optimization.rs
use anyhow::Result;
use std::collections::HashMap;
use std::sync::Arc;
use std::time::{Duration, Instant};
use tokio::sync::RwLock;
use tracing::{error, info, warn};

use crate::caching::{CacheManager, CacheStats};
use crate::compression::{CompressionManager, CompressionStats};
use crate::memory::{MemoryOptimizer, MemoryStats};
use crate::performance::{PerformanceOptimizer, PerformanceReport};
use crate::profiling::{Profiler, ProfilingReport};

/// Main optimization manager
pub struct OptimizationManager {
    memory_optimizer: MemoryOptimizer,
    performance_optimizer: PerformanceOptimizer,
    cache_manager: Arc<CacheManager>,
    compression_manager: Arc<RwLock<CompressionManager>>,
    profiler: Arc<Profiler>,
    optimization_rules: Vec<OptimizationRule>,
    last_optimization: Instant,
    optimization_interval: Duration,
}

#[derive(Debug, Clone)]
pub struct OptimizationRule {
    pub name: String,
    pub condition: OptimizationCondition,
    pub action: OptimizationAction,
    pub enabled: bool,
    pub priority: u32,
}

#[derive(Debug, Clone)]
pub enum OptimizationCondition {
    MemoryUsageAbove(usize),
    CpuUsageAbove(f64),
    CacheHitRateBelow(f64),
    CompressionRatioBelow(f64),
    LatencyAbove(Duration),
    ThroughputBelow(f64),
    ErrorRateAbove(f64),
    Custom(Box<dyn Fn(&OptimizationContext) -> bool + Send + Sync>),
}

#[derive(Debug, Clone)]
pub enum OptimizationAction {
    EnableMemoryCompression,
    IncreaseCacheSize,
    EnableDataCompression,
    ReduceConcurrency,
    ForceGarbageCollection,
    OptimizeDataStructures,
    EnableCaching,
    DisableProfiling,
    ScaleUp,
    ScaleDown,
    Custom(Box<dyn Fn(&mut OptimizationContext) -> Result<()> + Send + Sync>),
}

#[derive(Debug, Clone)]
pub struct OptimizationContext {
    pub memory_stats: MemoryStats,
    pub performance_report: PerformanceReport,
    pub cache_stats: HashMap<String, CacheStats>,
    pub compression_stats: CompressionStats,
    pub profiling_report: ProfilingReport,
    pub system_metrics: SystemMetrics,
}

#[derive(Debug, Clone)]
pub struct SystemMetrics {
    pub cpu_usage: f64,
    pub memory_usage: usize,
    pub disk_usage: usize,
    pub network_usage: usize,
    pub uptime: Duration,
}

impl OptimizationManager {
    pub fn new() -> Result<Self> {
        Ok(Self {
            memory_optimizer: MemoryOptimizer::new()?,
            performance_optimizer: PerformanceOptimizer::new(),
            cache_manager: Arc::new(CacheManager::new()),
            compression_manager: Arc::new(RwLock::new(CompressionManager::new(Default::default()))),
            profiler: Arc::new(Profiler::new()),
            optimization_rules: Vec::new(),
            last_optimization: Instant::now(),
            optimization_interval: Duration::from_secs(30),
        })
    }

    pub fn add_optimization_rule(&mut self, rule: OptimizationRule) {
        self.optimization_rules.push(rule);
        self.optimization_rules.sort_by_key(|r| r.priority);
    }

    pub async fn run_optimization_cycle(&mut self) -> Result<OptimizationResult> {
        let start = Instant::now();
        let mut applied_optimizations = Vec::new();
        let mut errors = Vec::new();

        // Collect current metrics
        let context = self.collect_context().await?;

        // Evaluate and apply optimization rules
        for rule in &self.optimization_rules {
            if !rule.enabled {
                continue;
            }

            match self.evaluate_condition(&rule.condition, &context).await {
                Ok(true) => match self.apply_action(&rule.action, &mut context).await {
                    Ok(()) => {
                        applied_optimizations.push(rule.name.clone());
                        info!("Applied optimization: {}", rule.name);
                    }
                    Err(e) => {
                        errors.push(format!("Failed to apply {}: {}", rule.name, e));
                    }
                },
                Ok(false) => {
                    // Condition not met, continue
                }
                Err(e) => {
                    errors.push(format!(
                        "Failed to evaluate condition for {}: {}",
                        rule.name, e
                    ));
                }
            }
        }

        // Run automatic optimizations
        self.run_automatic_optimizations().await?;

        self.last_optimization = Instant::now();

        Ok(OptimizationResult {
            applied_optimizations,
            errors,
            duration: start.elapsed(),
            context,
        })
    }

    async fn collect_context(&self) -> Result<OptimizationContext> {
        let memory_stats = MemoryStats::new(); // Placeholder
        let performance_report = self.performance_optimizer.get_performance_report().await;
        let cache_stats = self.cache_manager.get_all_stats().await;
        let compression_stats = self.compression_manager.read().await.get_stats().clone();
        let profiling_report = self.profiler.generate_report().await;
        let system_metrics = SystemMetrics {
            cpu_usage: 0.0,                 // Placeholder
            memory_usage: 0,                // Placeholder
            disk_usage: 0,                  // Placeholder
            network_usage: 0,               // Placeholder
            uptime: Duration::from_secs(0), // Placeholder
        };

        Ok(OptimizationContext {
            memory_stats,
            performance_report,
            cache_stats,
            compression_stats,
            profiling_report,
            system_metrics,
        })
    }

    async fn evaluate_condition(
        &self,
        condition: &OptimizationCondition,
        context: &OptimizationContext,
    ) -> Result<bool> {
        match condition {
            OptimizationCondition::MemoryUsageAbove(threshold) => {
                Ok(context.memory_stats.current_usage > *threshold)
            }
            OptimizationCondition::CpuUsageAbove(threshold) => {
                Ok(context.system_metrics.cpu_usage > *threshold)
            }
            OptimizationCondition::CacheHitRateBelow(threshold) => {
                // Calculate overall cache hit rate
                let total_hits: usize = context.cache_stats.values().map(|s| s.l1_entries).sum();
                let total_requests: usize =
                    context.cache_stats.values().map(|s| s.total_entries).sum();
                let hit_rate = if total_requests > 0 {
                    total_hits as f64 / total_requests as f64
                } else {
                    1.0
                };
                Ok(hit_rate < *threshold)
            }
            OptimizationCondition::CompressionRatioBelow(threshold) => {
                let ratio = context.compression_stats.total_compressed as f64
                    / context.compression_stats.total_original as f64;
                Ok(ratio < *threshold)
            }
            OptimizationCondition::LatencyAbove(threshold) => {
                // Check if any operation has high latency
                context
                    .performance_report
                    .operations
                    .values()
                    .any(|op| op.avg_duration > *threshold)
            }
            OptimizationCondition::ThroughputBelow(threshold) => {
                // Calculate overall throughput
                let total_ops = context.performance_report.total_operations;
                let total_time = context.performance_report.total_time.as_secs_f64();
                let throughput = if total_time > 0.0 {
                    total_ops as f64 / total_time
                } else {
                    0.0
                };
                Ok(throughput < *threshold)
            }
            OptimizationCondition::ErrorRateAbove(threshold) => {
                // Calculate overall error rate
                let total_errors: u64 = context
                    .performance_report
                    .operations
                    .values()
                    .map(|op| op.error_count)
                    .sum();
                let total_ops = context.performance_report.total_operations;
                let error_rate = if total_ops > 0 {
                    total_errors as f64 / total_ops as f64
                } else {
                    0.0
                };
                Ok(error_rate > *threshold)
            }
            OptimizationCondition::Custom(f) => Ok(f(context)),
        }
    }

    async fn apply_action(
        &self,
        action: &OptimizationAction,
        context: &mut OptimizationContext,
    ) -> Result<()> {
        match action {
            OptimizationAction::EnableMemoryCompression => {
                self.memory_optimizer.enable_compression(true);
                info!("Enabled memory compression");
            }
            OptimizationAction::IncreaseCacheSize => {
                // This would increase cache sizes
                info!("Increased cache size");
            }
            OptimizationAction::EnableDataCompression => {
                // This would enable data compression
                info!("Enabled data compression");
            }
            OptimizationAction::ReduceConcurrency => {
                // This would reduce concurrency levels
                info!("Reduced concurrency");
            }
            OptimizationAction::ForceGarbageCollection => {
                self.memory_optimizer.optimize_memory()?;
                info!("Forced garbage collection");
            }
            OptimizationAction::OptimizeDataStructures => {
                // This would optimize data structures
                info!("Optimized data structures");
            }
            OptimizationAction::EnableCaching => {
                // This would enable caching
                info!("Enabled caching");
            }
            OptimizationAction::DisableProfiling => {
                // This would disable profiling to reduce overhead
                info!("Disabled profiling");
            }
            OptimizationAction::ScaleUp => {
                // This would scale up resources
                info!("Scaled up resources");
            }
            OptimizationAction::ScaleDown => {
                // This would scale down resources
                info!("Scaled down resources");
            }
            OptimizationAction::Custom(f) => {
                f(context)?;
            }
        }
        Ok(())
    }

    async fn run_automatic_optimizations(&self) -> Result<()> {
        // Automatic memory optimization
        if self.memory_optimizer.should_garbage_collect() {
            self.memory_optimizer.optimize_memory()?;
        }

        // Automatic cache cleanup
        self.cache_manager.cleanup_all().await;

        Ok(())
    }

    pub async fn get_optimization_status(&self) -> OptimizationStatus {
        let context = self
            .collect_context()
            .await
            .unwrap_or_else(|_| OptimizationContext {
                memory_stats: MemoryStats::new(),
                performance_report: PerformanceReport {
                    uptime: Duration::ZERO,
                    total_operations: 0,
                    cpu_usage: 0.0,
                    memory_usage: 0,
                    operations: HashMap::new(),
                },
                cache_stats: HashMap::new(),
                compression_stats: CompressionStats::default(),
                profiling_report: ProfilingReport {
                    profiles: HashMap::new(),
                    total_calls: 0,
                    total_time: Duration::ZERO,
                    generated_at: chrono::Utc::now(),
                },
                system_metrics: SystemMetrics {
                    cpu_usage: 0.0,
                    memory_usage: 0,
                    disk_usage: 0,
                    network_usage: 0,
                    uptime: Duration::ZERO,
                },
            });

        OptimizationStatus {
            memory_usage: context.memory_stats.current_usage,
            cpu_usage: context.system_metrics.cpu_usage,
            cache_hit_rate: self.calculate_cache_hit_rate(&context.cache_stats),
            compression_ratio: self.calculate_compression_ratio(&context.compression_stats),
            last_optimization: self.last_optimization,
            next_optimization: self.last_optimization + self.optimization_interval,
        }
    }

    fn calculate_cache_hit_rate(&self, cache_stats: &HashMap<String, CacheStats>) -> f64 {
        let total_hits: usize = cache_stats.values().map(|s| s.l1_entries).sum();
        let total_requests: usize = cache_stats.values().map(|s| s.total_entries).sum();

        if total_requests > 0 {
            total_hits as f64 / total_requests as f64
        } else {
            1.0
        }
    }

    fn calculate_compression_ratio(&self, compression_stats: &CompressionStats) -> f64 {
        if compression_stats.total_original > 0 {
            compression_stats.total_compressed as f64 / compression_stats.total_original as f64
        } else {
            1.0
        }
    }

    pub fn set_optimization_interval(&mut self, interval: Duration) {
        self.optimization_interval = interval;
    }

    pub async fn start_optimization_loop(&mut self) -> Result<()> {
        let mut interval = tokio::time::interval(self.optimization_interval);

        loop {
            interval.tick().await;

            match self.run_optimization_cycle().await {
                Ok(result) => {
                    if !result.applied_optimizations.is_empty() {
                        info!("Applied optimizations: {:?}", result.applied_optimizations);
                    }
                    if !result.errors.is_empty() {
                        for error in result.errors {
                            error!("Optimization error: {}", error);
                        }
                    }
                }
                Err(e) => {
                    error!("Optimization cycle failed: {}", e);
                }
            }
        }
    }
}

#[derive(Debug, Clone)]
pub struct OptimizationResult {
    pub applied_optimizations: Vec<String>,
    pub errors: Vec<String>,
    pub duration: Duration,
    pub context: OptimizationContext,
}

#[derive(Debug, Clone)]
pub struct OptimizationStatus {
    pub memory_usage: usize,
    pub cpu_usage: f64,
    pub cache_hit_rate: f64,
    pub compression_ratio: f64,
    pub last_optimization: Instant,
    pub next_optimization: Instant,
}

/// Optimization recommendations
pub struct OptimizationRecommender {
    rules: Vec<RecommendationRule>,
}

#[derive(Debug, Clone)]
pub struct RecommendationRule {
    pub name: String,
    pub condition: Box<dyn Fn(&OptimizationContext) -> bool + Send + Sync>,
    pub recommendation: String,
    pub priority: u32,
}

impl OptimizationRecommender {
    pub fn new() -> Self {
        Self { rules: Vec::new() }
    }

    pub fn add_rule(&mut self, rule: RecommendationRule) {
        self.rules.push(rule);
        self.rules.sort_by_key(|r| r.priority);
    }

    pub fn get_recommendations(&self, context: &OptimizationContext) -> Vec<Recommendation> {
        let mut recommendations = Vec::new();

        for rule in &self.rules {
            if rule.condition(context) {
                recommendations.push(Recommendation {
                    name: rule.name.clone(),
                    description: rule.recommendation.clone(),
                    priority: rule.priority,
                });
            }
        }

        recommendations.sort_by_key(|r| r.priority);
        recommendations
    }
}

#[derive(Debug, Clone)]
pub struct Recommendation {
    pub name: String,
    pub description: String,
    pub priority: u32,
}

/// Performance tuning utilities
pub struct PerformanceTuner {
    optimization_manager: OptimizationManager,
    recommender: OptimizationRecommender,
}

impl PerformanceTuner {
    pub fn new() -> Result<Self> {
        Ok(Self {
            optimization_manager: OptimizationManager::new()?,
            recommender: OptimizationRecommender::new(),
        })
    }

    pub async fn tune_performance(&mut self) -> Result<TuningResult> {
        let context = self.optimization_manager.collect_context().await?;
        let recommendations = self.recommender.get_recommendations(&context);
        let optimization_result = self.optimization_manager.run_optimization_cycle().await?;

        Ok(TuningResult {
            recommendations,
            optimization_result,
            performance_improvement: self.calculate_improvement(&context),
        })
    }

    fn calculate_improvement(&self, context: &OptimizationContext) -> PerformanceImprovement {
        // Calculate performance improvement metrics
        PerformanceImprovement {
            memory_savings: 0,                 // Placeholder
            cpu_reduction: 0.0,                // Placeholder
            latency_reduction: Duration::ZERO, // Placeholder
            throughput_increase: 0.0,          // Placeholder
        }
    }
}

#[derive(Debug, Clone)]
pub struct TuningResult {
    pub recommendations: Vec<Recommendation>,
    pub optimization_result: OptimizationResult,
    pub performance_improvement: PerformanceImprovement,
}

#[derive(Debug, Clone)]
pub struct PerformanceImprovement {
    pub memory_savings: usize,
    pub cpu_reduction: f64,
    pub latency_reduction: Duration,
    pub throughput_increase: f64,
}

#[cfg(test)]
mod tests {
    use super::*;
    use std::time::Duration;

    #[tokio::test]
    async fn test_optimization_manager() {
        let mut manager = OptimizationManager::new().unwrap();

        let rule = OptimizationRule {
            name: "test_rule".to_string(),
            condition: OptimizationCondition::MemoryUsageAbove(1000),
            action: OptimizationAction::ForceGarbageCollection,
            enabled: true,
            priority: 1,
        };

        manager.add_optimization_rule(rule);

        let result = manager.run_optimization_cycle().await.unwrap();
        assert!(
            result.applied_optimizations.is_empty()
                || result
                    .applied_optimizations
                    .contains(&"test_rule".to_string())
        );
    }

    #[tokio::test]
    async fn test_optimization_recommender() {
        let mut recommender = OptimizationRecommender::new();

        let rule = RecommendationRule {
            name: "test_recommendation".to_string(),
            condition: Box::new(|_| true),
            recommendation: "Test recommendation".to_string(),
            priority: 1,
        };

        recommender.add_rule(rule);

        let context = OptimizationContext {
            memory_stats: MemoryStats::new(),
            performance_report: PerformanceReport {
                uptime: Duration::ZERO,
                total_operations: 0,
                cpu_usage: 0.0,
                memory_usage: 0,
                operations: HashMap::new(),
            },
            cache_stats: HashMap::new(),
            compression_stats: CompressionStats::default(),
            profiling_report: ProfilingReport {
                profiles: HashMap::new(),
                total_calls: 0,
                total_time: Duration::ZERO,
                generated_at: chrono::Utc::now(),
            },
            system_metrics: SystemMetrics {
                cpu_usage: 0.0,
                memory_usage: 0,
                disk_usage: 0,
                network_usage: 0,
                uptime: Duration::ZERO,
            },
        };

        let recommendations = recommender.get_recommendations(&context);
        assert!(!recommendations.is_empty());
        assert_eq!(recommendations[0].name, "test_recommendation");
    }
}
